baseline/llava/requirements.txt
transformers==4.40.2
torch==2.3.0
pillow==10.3.0
pandas==2.2.2
diffusers==0.27.2
matplotlib==3.8.4
accelerate==0.30.1
openai==1.25.0
python-dotenv==0.21.0
